+++ merlin.core/src/Merlin/Core/Timestep.h
@@ -5,17 +5,17 @@ namespace Merlin {
 	class Timestep
 	{
 	public:
-		Timestep(float time = 0.0f)
+		Timestep(double time = 0.0f)
 			: m_Time(time)
 		{
 		}
 
 		operator float() const { return m_Time; }
 
-		float GetSeconds() const { return m_Time; }
-		float GetMilliseconds() const { return m_Time * 1000.0f; }
+		double GetSeconds() const { return m_Time; }
+		double GetMilliseconds() const { return m_Time * 1000.0f; }
 	private:
-		float m_Time;
+		double m_Time;
 	};
 
 }
